requirements.txt
asgiref==3.4.1
astroid==2.6.5
atomicwrites==1.4.0
attrs==21.2.0
autoflake==1.4
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
dataclasses==0.6
Django==2.2
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djongo==1.2.31
Faker==8.11.0
idna==3.2
iniconfig==1.1.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mixer==7.1.2
openapi-codec==1.3.2
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.6
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
